requirements.txt
blessings==1.6
bpython==0.14.2
curtsies==0.1.19
gitdb==0.6.4
GitPython==1.0.1
greenlet==0.4.9
Pygments==2.0.2
requests==2.8.1
six==1.10.0
smmap==0.9.0
wheel==0.24.0
